Add files via upload
My coursework for Python at QMUL involves creating a Jupyter Notebook that adheres to strict guidelines and requirements. The notebook must be written in Python 3 and is expected to run smoothly on QMUL's JupyterHub, passing all tests when executed sequentially on a fresh kernel. It should group all import statements in one cell, include only the final answer code, and ensure my full name and student number are present. The project forbids the use of Biopython and modifications to the provided assignment and test code. The submission must be made in a .zip or .gz format on QMPlus, with all tests successfully passing to achieve full marks.